emailag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emailag"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "ulid-py",
    "python-dotenv",
    "aiohttp",
    "tenacity",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "coverage",
    "black",
    "ruff",
    "httpx",
]

[project.scripts]
agcom-api = "agcom_api.main:run"
agcom = "agcom.console.cli:main"
agcom-viewer = "agcom_viewer.main:run"

[tool.hatch.build.targets.wheel]
packages = ["agcom", "agcom_api", "agcom_viewer"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
